refactor: Ui 폴더 이름 -> UiLogic 으로 변경
